Huge changes for jhttpd, minor Vstr changes
 Vstr
  Tweak some documentation.

	* include/vstr-switch.h: Added VSTR_COMPILE_BUILTINS switch.
	(VSTR__AT_COMPILE_STRLEN): Added code to use __builtin_strlen() in GCC.

	* src/vstr.c (vstr_init): Factor the malloc-check APIs out into
	malloc-check.h

	* include/vstr-inline.h (vstr_wrap_memcpy): Do more "inline".
	(vstr_wrap_memcmp): Do more "inline".
	(vstr_wrap_memchr): Do more "inline".
	(vstr_wrap_memset): Do more "inline".

 mk.h
  Helper for examples to use malloc-check.h

 cntl
  Convert to mk allocations.
  Run a scan of the close queue after creating children.
  Allow children to optionly die. via. PDEATHSIG.

 conf
  Allow configurations to stay around.
  Allow user defined types.
  Make escaped and unescaped strings different types.
  Make "\\\n" parsing at the begining of strings not escaped.
  Minor code tidy in lexing, big code tidy for inlines.
  Error if data still left to be parsed.
  Speed up token retrieval.
  Fix some parse helper functions that didn't die on no values.
  Tweaked backtrace code.
  Added a bunch of functions.
  
 date
  Used <foo>_r time functions ... so hopefully they'll work from signals.

 evnt
  Moved child creation/death to be "events".
  Speedup via. calling blocking accept(), if we can (no "thundering herd").
  Use memcpy() for time/gettimeofday when in a single loop.
  Changed accept callback function to give the event we are processing.
   For allowing multiple listeners.
  Use mk for allocations.
  Bunch of bug fixes.
  Allow timeouts to work on io_w_shutdown evnts.
  Shutdown_w on timeouts ... and then re-timeout.
  Work around another timer_q bug.

 ex_cntl
  Use mk for allocations
  Handle NULL timeout cb's per new evnt code.

 ex_conf
  Use mk for allocations
  Print out in a prettier format.

 ex_dns
  Use mk for allocations
  Handle NULL timeout cb's per new evnt code.

 ex_echod
  Use mk for allocations
  Handle NULL timeout cb's per new evnt code.
  Bare minimum port of config. changes for httpd.

 ex_highlight
  Remove force inline off.

 ex_httpd
  Tweak main httpd documentation.
  Fix CFLAGS in automake.
  Move most of code to httpd.c
  Use mk for allocations
  Configuration for whether we check the host headers.
  Make signal logging "safer".

 httpd
  Moved code from ex_httpd, lots of changes.
  Configuration is now policy based, and allows policy changing based on ip addresses or
   the request beign made. Can also use this for denying connections.
  Code for multiple listeners is almost complete.

 ex_utils
  Added file output to ASSERT() calls (used from multi translation unit apps.)

 gzip-r.pl
  Don't create symlinks to non-existant files.

 httpd_policy
  Auto Vary headers.

 httpd_conf_main
  Make multiple config. files/arguments work.
  Policies.
  Vary headers

 httpd_conf_req
  Vary headers.
  Building paths for filename/Location, and path limiting.
  Allow persistent configurations.

 make_index.pl
  Default to css from www.and.org
  Output can be different to directory.
  Only use basename for title.

 mime_types
  Allow multiple configurations.
  Allow combining of configurations.

 opt
  Fix problem with resolving vuid/vgid NULL.

 opt_serv
  Tidy code using helpers.
  Add username/groupname that resolve.
  Add parent-death-signal toggle.
  Add (rlimit (file-num ...))

 vlg
  Create "signal safe" logging functions.

 Major unit test changes to go with above work.

git-archimport-id: james@and.org--2004-code/vstr--main--1--patch-109

+++ examples/gzip-r.pl
@@ -62,11 +62,12 @@ sub gzip_file
 
 	defined($dst) || die "Can't readlink $name: $!";
 
-	if ($dst !~ /$filter_exts_re/)
+	my $dst_gz = $dst . ".gz";
+	if (($dst !~ /$filter_exts_re/) && -f $dst_gz)
 	  {
 	    unlink($namegz);
 	    print STDOUT "Symlink: $name => $dst\n" if ($verbose_compress > 1);
-	    symlink($dst . ".gz", $namegz) || die "Can't symlink($namegz): $!";;
+	    symlink($dst_gz, $namegz) || die "Can't symlink($namegz): $!";
 	  }
 	return;
       }
